TCP looks complete. This change also includes fixups for to_i across several protocols (properties return expected integer values, instead of StructFu objects now). This needs to be double-checked across all protocols.
UDP's checksum on altered packets is still wrong. Need to chase that down. It's doubly
mysterious because TCP's checksum seems ok now.

Also found a bug where IP header lengths weren't being generated correctly out of
the gate (needed a recalc to become correct). This was irritating, so fixed that and
also gave a more sane default starting value for TTL so Wireshark won't complain.

This is a pretty major milestone here, so I expect, once I get UDP chased down, I'll
be able to verify that all the existing example scripts work with the new struct version.
Once that bit of QA is done, this looks ready for a proper release.



git-svn-id: https://packetfu.googlecode.com/svn/branches/struct_version@114 ebf01b17-4b54-0410-ae1b-bdaaa366248d