Incrementing version to 1.1.6
This is a pretty significant change from 1.1.5 and prior internally,
since the file structure for protocol definitions has changed, and most
functions no longer rely on method_missing. However, there are
effectively no API changes, so it's merely a patch-level change (as
opposed to a minor or major version change).